Move RealmCache to its own file
