+++ FeedStoreChallenge/RealmFeedStore.swift
@@ -9,27 +9,6 @@
 import Foundation
 import RealmSwift
 
-internal class RealmCache: Object {
-	@objc dynamic var _id: String!
-	@objc dynamic var timestamp: Date = Date()
-	let feed: List<RealmFeedImage> = List<RealmFeedImage>()
-	
-	internal convenience init(_id: String, feed: [RealmFeedImage], timestamp: Date) {
-		self.init()
-		self._id = _id
-		self.feed.append(objectsIn: feed)
-		self.timestamp = timestamp
-	}
-	
-	override class func primaryKey() -> String? {
-		return "_id"
-	}
-	
-	internal var local: [LocalFeedImage] {
-		return feed.map { $0.local }
-	}
-}
-
 internal class RealmFeedImage: Object {
 	@objc dynamic var _id: String!
 	@objc dynamic var desc: String?
